backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend API for Secret Santa App"
authors = ["Timur Smirnov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic = "^2.10.1"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
greenlet = "^3.1.1"
python-jose = "^3.3.0"
pyjwt = "^2.10.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.17"
websockets = "^14.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
